serving_pipeline/streamlit-tutorial/requirements.txt
streamlit
numpy
pandas
scikit-learn
torch
torchvision
opencv-python
Pillow
matplotlib
pyyaml
tqdm
scipy
